requirements.txt
aiosqlite
discord.py[voice]
asyncio
aiohttp